fix: move logic from user and payment to paymentcontainer
